fixed scheduling in cleaner way
